+++ bin/faucet/src/api/get_tokens.rs
@@ -11,7 +11,7 @@ use miden_faucet_lib::requests::{
     MintRequest,
     MintRequestSender,
 };
-use miden_faucet_lib::types::{AssetAmount, AssetAmountError, NoteType};
+use miden_faucet_lib::types::{AssetAmount, AssetAmountError};
 use miden_pow_rate_limiter::ChallengeError;
 use tokio::sync::mpsc::error::TrySendError;
 use tokio::sync::oneshot;
@@ -28,7 +28,6 @@ use crate::api_key::ApiKey;
     parent = None, target = COMPONENT, name = "server.get_tokens", skip_all, err,
     fields(
         account_id = %request.account_id,
-        is_private_note = %request.is_private_note,
         asset_amount = %request.asset_amount,
     )
 )]
@@ -96,6 +95,8 @@ pub enum MintRequestError {
     PowError(#[from] ChallengeError),
     #[error("API key {0} is invalid")]
     InvalidApiKey(String),
+    #[error("private notes are not supported")]
+    PrivateNoteUnsupported,
 }
 
 #[derive(Debug, thiserror::Error)]
@@ -188,6 +189,7 @@ impl IntoResponse for GetTokenError {
 /// # Errors
 ///
 /// Returns an error if:
+///   - a private note was requested
 ///   - the account ID is not a valid hex string
 ///   - the asset amount is not one of the provided options
 ///   - the API key is invalid
@@ -199,11 +201,9 @@ fn validate_get_tokens_params(
     params: &GetTokensQueryParams,
     server: &ApiServer,
 ) -> Result<MintRequest, MintRequestError> {
-    let note_type = if params.is_private_note {
-        NoteType::Private
-    } else {
-        NoteType::Public
-    };
+    if params.is_private_note == Some(true) {
+        return Err(MintRequestError::PrivateNoteUnsupported);
+    }
 
     let account_id = if params.account_id.starts_with("0x") {
         AccountId::from_hex(&params.account_id).map_err(AccountError::ParseId)
@@ -245,5 +245,5 @@ fn validate_get_tokens_params(
         request_complexity,
     )?;
 
-    Ok(MintRequest { account_id, note_type, asset_amount })
+    Ok(MintRequest { account_id, asset_amount })
 }