+++ crates/faucet/src/types.rs
@@ -60,28 +60,3 @@ pub enum AssetAmountError {
       )]
     AssetAmountTooBig(u64),
 }
-
-/// Type of note to generate for a mint request.
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub enum NoteType {
-    Private,
-    Public,
-}
-
-impl From<NoteType> for miden_client::note::NoteType {
-    fn from(value: NoteType) -> Self {
-        match value {
-            NoteType::Private => Self::Private,
-            NoteType::Public => Self::Public,
-        }
-    }
-}
-
-impl std::fmt::Display for NoteType {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Private => f.write_str("private"),
-            Self::Public => f.write_str("public"),
-        }
-    }
-}