+++ docs/src/examples/typescript/request_tokens.ts
@@ -1,6 +1,5 @@
 import { sha256 } from '@noble/hashes/sha2.js';
 import { hexToBytes } from '@noble/hashes/utils.js';
-import fs from 'fs';
 
 async function sendPowRequest(baseUrl: string, accountId: string): Promise<{ challenge: string, target: bigint }> {
     const powUrl = new URL('/pow', baseUrl);
@@ -50,7 +49,6 @@ async function solveChallenge(challenge: string, target: bigint): Promise<number
 async function getTokens(baseUrl: string, account_id: string, nonce: number, challenge: string): Promise<{ noteId: string, txId: string }> {
     const params = new URLSearchParams({
         account_id: account_id,
-        is_private_note: 'true',
         asset_amount: '100',
         challenge: challenge,
         nonce: nonce.toString()
@@ -66,20 +64,6 @@ async function getTokens(baseUrl: string, account_id: string, nonce: number, cha
     return { noteId, txId };
 }
 
-async function downloadNote(baseUrl: string, noteId: string): Promise<void> {
-    const url = `${baseUrl}/get_note?note_id=${noteId}`;
-    const response = await fetch(url);
-    if (!response.ok) throw new Error(`Get note error: ${response.status} ${await response.text()}`);
-
-    const text = await response.text();
-    const json = JSON.parse(text);
-
-    // Decode note with base64
-    const noteData = Buffer.from(json.data_base64, 'base64');
-
-    fs.writeFileSync('note.mno', noteData);
-}
-
 async function main(): Promise<void> {
     const baseUrl = 'http://localhost:8000';
     const accountId = '0xca8203e8e58cf72049b061afca78ce';
@@ -89,7 +73,6 @@ async function main(): Promise<void> {
     let { noteId, txId } = await getTokens(baseUrl, accountId, nonce, challenge);
     console.log('Note ID:', noteId);
     console.log('Tx ID:', txId);
-    await downloadNote(baseUrl, noteId);
 }
 
 main();